+++ src/main/java/fi/dy/masa/malilib/network/IPluginClientPlayHandler.java
@@ -3,8 +3,8 @@ package fi.dy.masa.malilib.network;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Objects;
-import net.neoforged.neoforge.network.handling.IPayloadContext;
-import net.neoforged.neoforge.network.handling.IPayloadHandler;
+import org.thinkingstudio.fabric.api.client.networking.v1.ClientPlayNetworking;
+import org.thinkingstudio.fabric.api.networking.v1.PayloadTypeRegistry;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import net.minecraft.client.network.ClientPlayNetworkHandler;
 import net.minecraft.nbt.NbtCompound;
@@ -16,13 +16,12 @@ import net.minecraft.network.packet.Packet;
 import net.minecraft.network.packet.c2s.common.CustomPayloadC2SPacket;
 import net.minecraft.util.Identifier;
 import fi.dy.masa.malilib.MaLiLib;
-import org.thinkingstudio.mafglib.util.NeoNetwork;
 
 /**
  * Interface for ClientPlayHandler, for downstream mods.
  * @param <T> (Payload)
  */
-public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayloadHandler<T>
+public interface IPluginClientPlayHandler<T extends CustomPayload> extends ClientPlayNetworking.PlayPayloadHandler<T>
 {
     int FROM_SERVER = 1;
     int TO_SERVER = 2;
@@ -57,8 +56,8 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     void reset(Identifier channel);
 
     /**
-     * Register your Payload with NeoForge network.
-     * See the NeoForge's network Java Docs under PayloadTypeRegistry -> register()
+     * Register your Payload with Fabric API.
+     * See the fabric-networking-api-v1 Java Docs under PayloadTypeRegistry -> register()
      * for more information on how to do this.
      * -
      * @param id (Your Payload Id<T>)
@@ -73,9 +72,13 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
             {
                 switch (direction)
                 {
-                    case TO_SERVER, FROM_CLIENT -> NeoNetwork.playToServer(id, codec, this);
-                    case FROM_SERVER, TO_CLIENT -> NeoNetwork.playToClient(id, codec, this);
-                    default -> NeoNetwork.playBidirectional(id, codec, this);
+                    case TO_SERVER, FROM_CLIENT -> PayloadTypeRegistry.playC2S().register(id, codec);
+                    case FROM_SERVER, TO_CLIENT -> PayloadTypeRegistry.playS2C().register(id, codec);
+                    default ->
+                    {
+                        PayloadTypeRegistry.playC2S().register(id, codec);
+                        PayloadTypeRegistry.playS2C().register(id, codec);
+                    }
                 }
             }
             catch (IllegalArgumentException e)
@@ -93,22 +96,20 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     /**
      * Register your Packet Receiver function.
      * You can use the HANDLER itself (Singleton method), or any other class that you choose.
-     * See the NeoForge's network Java Docs under PayloadRegistrar.playToClient()
+     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.registerGlobalReceiver()
      * for more information on how to do this.
      * -
      * @param id (Your Payload Id<T>)
-     * @param codec (Your Payload's CODEC)
      * @param receiver (Your Packet Receiver // if null, uses this::receivePlayPayload)
      * @return (True / False)
      */
-    default boolean registerPlayReceiver(@Nonnull CustomPayload.Id<T> id, @Nonnull PacketCodec<? super RegistryByteBuf,T> codec, @Nullable IPayloadHandler<T> receiver)
+    default boolean registerPlayReceiver(@Nonnull CustomPayload.Id<T> id, @Nullable ClientPlayNetworking.PlayPayloadHandler<T> receiver)
     {
         if (this.isPlayRegistered(this.getPayloadChannel()))
         {
             try
             {
-                NeoNetwork.playToClient(id, codec, Objects.requireNonNullElse(receiver, this::receivePlayPayload));
-                return true;
+                return ClientPlayNetworking.registerGlobalReceiver(id, Objects.requireNonNullElse(receiver, this::receivePlayPayload));
             }
             catch (IllegalArgumentException e)
             {
@@ -121,24 +122,24 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
         return false;
     }
 
-//    /**
-//     * Unregisters your Packet Receiver function.
-//     * You can use the HANDLER itself (Singleton method), or any other class that you choose.
-//     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.unregisterGlobalReceiver()
-//     * for more information on how to do this.
-//     */
-//    default void unregisterPlayReceiver()
-//    {
-//        ClientPlayNetworking.unregisterGlobalReceiver(this.getPayloadChannel());
-//    }
+    /**
+     * Unregisters your Packet Receiver function.
+     * You can use the HANDLER itself (Singleton method), or any other class that you choose.
+     * See the fabric-network-api-v1 Java Docs under ClientPlayNetworking.unregisterGlobalReceiver()
+     * for more information on how to do this.
+     */
+    default void unregisterPlayReceiver()
+    {
+        ClientPlayNetworking.unregisterGlobalReceiver(this.getPayloadChannel());
+    }
 
     /**
      * Receive Payload by pointing static receive() method to this to convert Payload to its data decode() function.
      * -
      * @param payload (Payload to decode)
-     * @param ctx (NeoForge's network Context)
+     * @param ctx (Fabric Context)
      */
-    void receivePlayPayload(T payload, IPayloadContext ctx);
+    void receivePlayPayload(T payload, ClientPlayNetworking.Context ctx);
 
     /**
      * Receive Payload via the legacy "onCustomPayload" from a Network Handler Mixin interface.
@@ -182,7 +183,7 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     void encodeWithSplitter(PacketByteBuf buf, ClientPlayNetworkHandler handler);
 
     /**
-     * Sends the Payload to the server using the NeoForge's network interface.
+     * Sends the Payload to the server using the Fabric-API interface.
      * -
      * @param payload (The Payload to send)
      * @return (true/false --> for error control)
@@ -191,15 +192,15 @@ public interface IPluginClientPlayHandler<T extends CustomPayload> extends IPayl
     {
         if (payload.getId().id().equals(this.getPayloadChannel()) && this.isPlayRegistered(this.getPayloadChannel()))
         {
-            if (NeoNetwork.canSend(payload.getId()))
+            if (ClientPlayNetworking.canSend(payload.getId()))
             {
-                NeoNetwork.sendToServer(payload);
+                ClientPlayNetworking.send(payload);
                 return true;
             }
         }
         else
         {
-            MaLiLib.logger.warn("sendPlayPayload: [NeoForge-Network] error sending payload for channel: {}, check if channel is registered", payload.getId().id().toString());
+            MaLiLib.logger.warn("sendPlayPayload: [Fabric-API] error sending payload for channel: {}, check if channel is registered", payload.getId().id().toString());
         }
 
         return false;